requirements.txt
fastapi>=0.116.0
openai>=1.0.0
python-dotenv>=1.0.1
uvicorn[standard]>=0.35.0
